+++ R/mod_edish.R
@@ -7,6 +7,7 @@ EDISH <- pack_of_constants(
   X_AXIS_HEADER = "Specify x-axis",
   Y_AXIS_HEADER = "Specify y-axis",
   X_AXIS_ID = "x_axis",
+  Y_AXIS_ID = "y_axis",
   AXIS_LABEL = "Parameter:",
   X_REF_ID = "x_ref",
   Y_REF_ID = "y_ref",
@@ -50,7 +51,8 @@ edish_UI <- function(module_id,
                      arm_default_vals,
                      at_choices,
                      at_default_val,
-                     tbili_choice,
+                     tbili_choices,
+                     tbili_default_val,
                      default_by_visit,
                      window_days) {
 
@@ -80,13 +82,6 @@ edish_UI <- function(module_id,
         6,
         shiny::div(
           shiny::h4(EDISH$X_AXIS_HEADER),
-          shiny::checkboxInput(
-            ns(EDISH$BY_VISIT_ID),
-            label = shiny::span(EDISH$BY_VISIT_LABEL,
-                                shiny::icon("circle-info",
-                                            title = EDISH$BY_VISIT_INFO)),
-            value = default_by_visit
-          ),
           shiny::selectInput(
             inputId = ns(EDISH$X_AXIS_ID),
             label = EDISH$AXIS_LABEL,
@@ -109,6 +104,13 @@ edish_UI <- function(module_id,
             max = 100,
             step = 0.1
           ),
+          shiny::checkboxInput(
+            ns(EDISH$BY_VISIT_ID),
+            label = shiny::span(EDISH$BY_VISIT_LABEL,
+                                shiny::icon("circle-info",
+                                            title = EDISH$BY_VISIT_INFO)),
+            value = default_by_visit
+          ),
           shiny::radioButtons(
             inputId = ns(EDISH$BASE_INCL_ID),
             label = EDISH$BASE_INCL_LABEL,
@@ -122,6 +124,12 @@ edish_UI <- function(module_id,
         6,
         shiny::div(
           shiny::h4(EDISH$Y_AXIS_HEADER),
+          shiny::selectInput(
+            inputId = ns(EDISH$Y_AXIS_ID),
+            label = EDISH$AXIS_LABEL,
+            choices = tbili_choices,
+            selected = tbili_default_val
+          ),
           shiny::numericInput(
             inputId = ns(EDISH$Y_REF_ID),
             label = EDISH$REF_LABEL,
@@ -190,15 +198,14 @@ edish_UI <- function(module_id,
 edish_server <- function(
     module_id,
     dataset_list,
+    lb_date_var,
     subjectid_var = "USUBJID",
     arm_var = "ACTARM",
     visit_var = "VISIT",
     baseline_visit_val = "VISIT 01",
     lb_test_var = "LBTEST",
     at_choices = NULL,
-    tbili_choice = NULL,
     alp_choice = NULL,
-    lb_date_var = NULL,
     lb_result_var = "LBSTRESN",
     ref_range_upper_lim_var = "LBSTNRHI",
     on_sbj_click = NULL) {
@@ -208,6 +215,7 @@ edish_server <- function(
   # Check validity of arguments
   ac <- checkmate::makeAssertCollection()
   checkmate::assert_multi_class(dataset_list, c("reactive", "shinymeta_reactive"), add = ac)
+  checkmate::assert_string(lb_date_var, min.chars = 1, add = ac)
   checkmate::assert_string(subjectid_var, min.chars = 1, add = ac)
   checkmate::assert_string(arm_var, min.chars = 1, add = ac)
   checkmate::assert_string(visit_var, min.chars = 1, add = ac)
@@ -215,9 +223,7 @@ edish_server <- function(
   checkmate::assert_string(lb_test_var, min.chars = 1, add = ac)
   checkmate::assert_character(at_choices, min.chars = 1, any.missing = FALSE, all.missing = FALSE,
                               unique = TRUE, min.len = 1, add = ac)
-  checkmate::assert_string(tbili_choice, min.chars = 1, add = ac)
   checkmate::assert_string(alp_choice, min.chars = 1, null.ok = TRUE, add = ac)
-  checkmate::assert_string(lb_date_var, min.chars = 1, add = ac)
   checkmate::assert_string(lb_result_var, min.chars = 1, add = ac)
   checkmate::assert_string(ref_range_upper_lim_var, min.chars = 1, add = ac)
   checkmate::reportAssertions(ac)
@@ -268,7 +274,7 @@ edish_server <- function(
         visit_var = visit_var,
         lb_test_var = lb_test_var,
         at_choices = at_choices,
-        tbili_choice = tbili_choice,
+        tbili_choice = input[[EDISH$Y_AXIS_ID]],
         alp_choice = alp_choice,
         lb_date_var = lb_date_var,
         lb_result_var = lb_result_var,
@@ -286,7 +292,7 @@ edish_server <- function(
           baseline_visit_val = baseline_visit_val,
           lb_test_var = lb_test_var,
           at_choices = at_choices,
-          tbili_choice = tbili_choice,
+          tbili_choice = input[[EDISH$Y_AXIS_ID]],
           norm_ref_type = nrt,
           alp_choice = alp_choice,
           lb_date_var = lb_date_var,
@@ -371,7 +377,7 @@ edish_server <- function(
         subjectid_var = subjectid_var,
         arm_var = arm_var,
         sel_x = input[[EDISH$X_AXIS_ID]],
-        sel_y = tbili_choice,
+        sel_y = input[[EDISH$Y_AXIS_ID]],
         norm_ref_type = input[[EDISH$PLOT_TYPE_ID]],
         x_ref_line_num = input[[EDISH$X_REF_ID]],
         y_ref_line_num = input[[EDISH$Y_REF_ID]],
@@ -416,7 +422,7 @@ edish_server <- function(
 #'
 #' `mod_edish()` displays the (modified) evaluation of Drug-Induced Serious Hepatotoxicity (eDISH/mDISH) plot
 #' to support the assessment of drug-induced liver injury (DILI). The scatter plot depicts the correlation
-#' between peak values of an aminotransferase parameter and total bilirubin on a subject level.
+#' between peak values of an aminotransferase parameter and a liver function parameter on a subject level.
 #'
 #' @param module_id `[character(1)]`
 #'
@@ -429,6 +435,10 @@ edish_server <- function(
 #' @param lab_dataset_name `[character(1)]`
 #'
 #' Name of the laboratory results dataset to be used.
+#'
+#' @param lb_date_var `[character(1)]`
+#'
+#' Name of the variable (`Date` or `POSIXt` class) containing the laboratory test date.
 #'
 #' @param subjectid_var `[character(1)]`
 #'
@@ -458,21 +468,22 @@ edish_server <- function(
 #'
 #' Character vector specifying the possible choices of the x-axis aminotransferase laboratory test.
 #'
-#' @param at_default_val `[character(1)]`
+#' @param at_default_val `[character(1) | NULL]`
 #'
 #' Character specifying the default x-axis aminotransferase laboratory test choice.
 #'
-#' @param tbili_choice `[character(1)]`
+#' @param tbili_choices `[character(1+)]`
 #'
-#' Character vector specifying the y-axis total bilirubin laboratory test choice.
+#' Character vector specifying the possible choices of the y-axis liver function laboratory test
+#' (e.g. "Bilirubin", "Prothrombin Intl. Normalized Ratio", etc.).
+#'
+#' @param tbili_default_val `[character(1) | NULL]`
+#'
+#' Character specifying the default y-axis liver function laboratory test choice.
 #'
 #' @param alp_choice `[character(1) | NULL]`
 #'
 #' Character vector specifying the alkaline phosphatase laboratory test choice.
-#'
-#' @param lb_date_var `[character(1)]`
-#'
-#' Name of the variable (`Date` or `POSIXt` class) containing the laboratory test date.
 #'
 #' @param lb_result_var `[character(1)]`
 #'
@@ -508,6 +519,7 @@ mod_edish <- function(
     module_id,
     subject_level_dataset_name,
     lab_dataset_name,
+    lb_date_var,
     subjectid_var = "USUBJID",
     arm_var = "ACTARM",
     arm_default_vals = NULL,
@@ -516,9 +528,9 @@ mod_edish <- function(
     lb_test_var = "LBTEST",
     at_choices = c("Alanine Aminotransferase", "Aspartate Aminotransferase"),
     at_default_val = "Aspartate Aminotransferase",
-    tbili_choice = "Bilirubin",
+    tbili_choices = "Bilirubin",
+    tbili_default_val = "Bilirubin",
     alp_choice = "Alkaline Phosphatase",
-    lb_date_var = NULL,
     lb_result_var = "LBSTRESN",
     ref_range_upper_lim_var = "LBSTNRHI",
     default_by_visit = FALSE,
@@ -531,7 +543,8 @@ mod_edish <- function(
                arm_default_vals = arm_default_vals,
                at_choices = at_choices,
                at_default_val = at_default_val,
-               tbili_choice = tbili_choice,
+               tbili_choices = tbili_choices,
+               tbili_default_val = tbili_default_val,
                default_by_visit = default_by_visit,
                window_days = window_days)
     },
@@ -557,15 +570,14 @@ mod_edish <- function(
       edish_server(
         module_id = module_id,
         dataset_list = dataset_list,
+        lb_date_var = lb_date_var,
         subjectid_var = subjectid_var,
         arm_var = arm_var,
         visit_var = visit_var,
         baseline_visit_val = baseline_visit_val,
         lb_test_var = lb_test_var,
         at_choices = at_choices,
-        tbili_choice = tbili_choice,
         alp_choice = alp_choice,
-        lb_date_var = lb_date_var,
         lb_result_var = lb_result_var,
         ref_range_upper_lim_var = ref_range_upper_lim_var,
         on_sbj_click = on_sbj_click_fun
@@ -584,6 +596,7 @@ mod_edish_API_docs <- list(
   module_id = list(""),
   subject_level_dataset_name = list(""),
   lab_dataset_name = list(""),
+  lb_date_var = list(""),
   subjectid_var = list(""),
   arm_var = list(""),
   arm_default_vals = list(""),
@@ -592,9 +605,9 @@ mod_edish_API_docs <- list(
   lb_test_var = list(""),
   at_choices = list(""),
   at_default_val = list(""),
-  tbili_choice = list(""),
+  tbili_choices = list(""),
+  tbili_default_val = list(""),
   alp_choice = list(""),
-  lb_date_var = list(""),
   lb_result_var = list(""),
   ref_range_upper_lim_var = list(""),
   default_by_visit = list(""),
@@ -606,6 +619,7 @@ mod_edish_API_spec <- TC$group(
   module_id = TC$mod_ID(),
   subject_level_dataset_name = TC$dataset_name() |> TC$flag("subject_level_dataset_name"),
   lab_dataset_name = TC$dataset_name(),
+  lb_date_var = TC$col("lab_dataset_name", TC$or(TC$date(), TC$datetime())),
   subjectid_var = TC$col("subject_level_dataset_name", TC$or(TC$character(), TC$factor())) |> TC$flag("subjid_var"),
   arm_var = TC$col("subject_level_dataset_name", TC$or(TC$character(), TC$factor())),
   arm_default_vals = TC$choice_from_col_contents("arm_var") |> TC$flag("one_or_more", "optional", "manual_check"),
@@ -614,9 +628,9 @@ mod_edish_API_spec <- TC$group(
   lb_test_var = TC$col("lab_dataset_name", TC$or(TC$character(), TC$factor())),
   at_choices = TC$choice_from_col_contents("lb_test_var") |> TC$flag("one_or_more"),
   at_default_val = TC$choice_from_col_contents("lb_test_var") |> TC$flag("optional"),
-  tbili_choice = TC$choice_from_col_contents("lb_test_var"),
+  tbili_choices = TC$choice_from_col_contents("lb_test_var") |> TC$flag("one_or_more"),
+  tbili_default_val = TC$choice_from_col_contents("lb_test_var") |> TC$flag("optional"),
   alp_choice = TC$choice_from_col_contents("lb_test_var") |> TC$flag("optional"),
-  lb_date_var = TC$col("lab_dataset_name", TC$or(TC$date(), TC$datetime())),
   lb_result_var = TC$col("lab_dataset_name", TC$numeric()),
   ref_range_upper_lim_var = TC$col("lab_dataset_name", TC$numeric()) |> TC$flag("optional"),
   default_by_visit = TC$logical() |> TC$flag("manual_check"),
@@ -625,20 +639,20 @@ mod_edish_API_spec <- TC$group(
 ) |> TC$attach_docs(mod_edish_API_docs)
 
 check_mod_edish <- function(
-    afmm, datasets, module_id, subject_level_dataset_name, lab_dataset_name, subjectid_var, arm_var, arm_default_vals,
-    visit_var, baseline_visit_val, lb_test_var,
-    at_choices, at_default_val, tbili_choice, alp_choice,
-    lb_date_var, lb_result_var, ref_range_upper_lim_var, default_by_visit, window_days, receiver_id
-  ) {
+    afmm, datasets, module_id, subject_level_dataset_name, lab_dataset_name, lb_date_var,
+    subjectid_var, arm_var, arm_default_vals, visit_var, baseline_visit_val, lb_test_var,
+    at_choices, at_default_val, tbili_choices, tbili_default_val, alp_choice, lb_result_var,
+    ref_range_upper_lim_var, default_by_visit, window_days, receiver_id
+) {
   warn <- CM$container()
   err <- CM$container()
 
   OK <- check_mod_edish_auto(
     afmm, datasets,
-    module_id, subject_level_dataset_name, lab_dataset_name, subjectid_var, arm_var, arm_default_vals,
-    visit_var, baseline_visit_val, lb_test_var,
-    at_choices, at_default_val, tbili_choice, alp_choice,
-    lb_date_var, lb_result_var, ref_range_upper_lim_var, window_days, default_by_visit, receiver_id,
+    module_id, subject_level_dataset_name, lab_dataset_name, lb_date_var,
+    subjectid_var, arm_var, arm_default_vals, visit_var, baseline_visit_val, lb_test_var,
+    at_choices, at_default_val, tbili_choices, tbili_default_val, alp_choice, lb_result_var,
+    ref_range_upper_lim_var, window_days, default_by_visit, receiver_id,
     warn, err
   )
 
@@ -686,12 +700,12 @@ check_mod_edish <- function(
       cond = !any(duplicated(all_vars)),
       msg = sprintf(
         "This modules expects the following variables to refer to unique columns:<br><pre>%s</pre>",
-        paste(capture.output(setNames(all_vars, var_parameters)), collapse = "\n")
+        paste(utils::capture.output(stats::setNames(all_vars, var_parameters)), collapse = "\n")
       )
     )
   }
 
-  # NOTE: Ensures that `lb_test_default_{x,y}_val` are a subset of the available `lb_test_choices`
+  # NOTE: Ensures that `at_default_val` is a subset of the available `at_choices`
   if (all(OK[c("lab_dataset_name", "lb_test_var", "at_choices", "at_default_val")])) {
     CM$assert(
       container = err,
@@ -699,6 +713,18 @@ check_mod_edish <- function(
       msg = sprintf(
         'The value assigned to `at_default_val` ("%s") should be among the ones provided by `at_choices` (%s).',
         at_default_val, paste(sprintf('"%s"', at_choices), collapse = ", ")
+      )
+    )
+  }
+
+  # NOTE: Ensures that `tbili_default_val` is a subset of the available `tbili_choices`
+  if (all(OK[c("lab_dataset_name", "lb_test_var", "tbili_choices", "tbili_default_val")])) {
+    CM$assert(
+      container = err,
+      cond = tbili_default_val %in% tbili_choices,
+      msg = sprintf(
+        'The value assigned to `tbili_default_val` ("%s") should be among the ones provided by `tbili_choices` (%s).',
+        tbili_default_val, paste(sprintf('"%s"', tbili_choices), collapse = ", ")
       )
     )
   }